testdaq function moved to rundaq
in windows even the buffer manager module is not legible to use as an argument to Process, so this is the version, where buffer manager is not passed as argument to process, but it is used as global variable. but that then requires testdaq functions to be imported on the main script for various reasons. at this level of changes, it is not good idea to push changes to original quast repo, but this needs many more tests and foreseeing